chore(content-manager): convert cm app shell to TS
