+++ packages/core/admin/admin/src/components/DragLayer.tsx
@@ -21,7 +21,10 @@ function getStyle(
 
 export interface DragLayerProps {
   renderItem: (item: {
-    item: unknown;
+    /**
+     * TODO: it'd be great if we could make this a union where the type infers the item.
+     */
+    item: any;
     type: ReturnType<DragLayerMonitor['getItemType']>;
   }) => React.ReactNode;
 }
