--- packages/core/admin/admin/src/content-manager/contexts/ModelsContext.jsx
@@ -1,5 +1,0 @@
-import { createContext } from 'react';
-
-const ModelsContext = createContext();
-
-export default ModelsContext;
