+++ packages/core/admin/admin/src/content-manager/contexts/models.ts
@@ -1,0 +1,9 @@
+import * as React from 'react';
+
+const ModelsContext = React.createContext<{
+  refetchData: () => Promise<void>;
+}>({
+  refetchData: () => Promise.resolve(),
+});
+
+export { ModelsContext };
