--- packages/core/admin/admin/src/content-manager/pages/App/index.jsx
@@ -1,169 +1,0 @@
-import React, { useEffect, useRef } from 'react';
-
-import { HeaderLayout, Layout, Main } from '@strapi/design-system';
-import {
-  AnErrorOccurred,
-  CheckPagePermissions,
-  LoadingIndicatorPage,
-  useGuidedTour,
-} from '@strapi/helper-plugin';
-import sortBy from 'lodash/sortBy';
-import { Helmet } from 'react-helmet';
-import { useIntl } from 'react-intl';
-import { useSelector } from 'react-redux';
-import { Redirect, Route, Switch, useLocation, useRouteMatch } from 'react-router-dom';
-
-import { DragLayer } from '../../../components/DragLayer';
-import { selectAdminPermissions } from '../../../selectors';
-import ModelsContext from '../../contexts/ModelsContext';
-import { ItemTypes } from '../../utils/dragAndDrop';
-import { getTranslation } from '../../utils/translations';
-import { CollectionTypeRecursivePath } from '../CollectionTypeRecursivePath';
-import ComponentSettingsView from '../ComponentSetttingsView';
-import NoContentType from '../NoContentType';
-import NoPermissions from '../NoPermissions';
-import { SingleTypeRecursivePath } from '../SingleTypeRecursivePath';
-
-import { CardDragPreview } from './components/CardDragPreview';
-import { ComponentDragPreview } from './components/ComponentDragPreview';
-import { RelationDragPreview } from './components/RelationDragPreview';
-import LeftMenu from './LeftMenu';
-import useContentManagerInitData from './useContentManagerInitData';
-
-function renderDraglayerItem({ type, item }) {
-  if ([ItemTypes.EDIT_FIELD, ItemTypes.FIELD].includes(type)) {
-    return <CardDragPreview labelField={item.labelField} />;
-  }
-
-  /**
-   * Because a user may have multiple relations / dynamic zones / repeable fields in the same content type,
-   * we append the fieldName for the item type to make them unique, however, we then want to extract that
-   * first type to apply the correct preview.
-   */
-  const [actualType] = type.split('_');
-
-  switch (actualType) {
-    case ItemTypes.COMPONENT:
-    case ItemTypes.DYNAMIC_ZONE:
-      return <ComponentDragPreview displayedValue={item.displayedValue} />;
-
-    case ItemTypes.RELATION:
-      return (
-        <RelationDragPreview
-          displayedValue={item.displayedValue}
-          status={item.status}
-          width={item.width}
-        />
-      );
-
-    default:
-      return null;
-  }
-}
-
-const App = () => {
-  const contentTypeMatch = useRouteMatch(`/content-manager/:kind/:uid`);
-  const { status, collectionTypeLinks, singleTypeLinks, models, refetchData } =
-    useContentManagerInitData();
-  const authorisedModels = sortBy([...collectionTypeLinks, ...singleTypeLinks], (model) =>
-    model.title.toLowerCase()
-  );
-  const { pathname } = useLocation();
-  const { formatMessage } = useIntl();
-  const { startSection } = useGuidedTour();
-  const startSectionRef = useRef(startSection);
-  const permissions = useSelector(selectAdminPermissions);
-
-  useEffect(() => {
-    if (startSectionRef.current) {
-      startSectionRef.current('contentManager');
-    }
-  }, []);
-
-  if (status === 'loading') {
-    return (
-      <Main aria-busy="true">
-        <HeaderLayout
-          title={formatMessage({
-            id: getTranslation('header.name'),
-            defaultMessage: 'Content',
-          })}
-        />
-        <LoadingIndicatorPage />
-      </Main>
-    );
-  }
-
-  // Array of models that are displayed in the content manager
-  const supportedModelsToDisplay = models.filter(({ isDisplayed }) => isDisplayed);
-
-  // Redirect the user to the 403 page
-  if (
-    authorisedModels.length === 0 &&
-    supportedModelsToDisplay.length > 0 &&
-    pathname !== '/content-manager/403'
-  ) {
-    return <Redirect to="/content-manager/403" />;
-  }
-
-  // Redirect the user to the create content type page
-  if (supportedModelsToDisplay.length === 0 && pathname !== '/content-manager/no-content-types') {
-    return <Redirect to="/content-manager/no-content-types" />;
-  }
-
-  if (!contentTypeMatch && authorisedModels.length > 0) {
-    return (
-      <Redirect
-        to={`${authorisedModels[0].to}${
-          authorisedModels[0].search ? `?${authorisedModels[0].search}` : ''
-        }`}
-      />
-    );
-  }
-
-  return (
-    <Layout sideNav={<LeftMenu />}>
-      <DragLayer renderItem={renderDraglayerItem} />
-      <ModelsContext.Provider value={{ refetchData }}>
-        <Switch>
-          <Route path="/content-manager/components/:uid/configurations/edit">
-            <CheckPagePermissions permissions={permissions.contentManager.componentsConfigurations}>
-              <ComponentSettingsView />
-            </CheckPagePermissions>
-          </Route>
-          <Route
-            path="/content-manager/collectionType/:slug"
-            component={CollectionTypeRecursivePath}
-          />
-          <Route path="/content-manager/singleType/:slug" component={SingleTypeRecursivePath} />
-
-          <Route path="/content-manager/403">
-            <NoPermissions />
-          </Route>
-          <Route path="/content-manager/no-content-types">
-            <NoContentType />
-          </Route>
-          <Route path="" component={AnErrorOccurred} />
-        </Switch>
-      </ModelsContext.Provider>
-    </Layout>
-  );
-};
-
-export { App };
-
-export default function () {
-  const { formatMessage } = useIntl();
-
-  return (
-    <>
-      <Helmet
-        title={formatMessage({
-          id: getTranslation('plugin.name'),
-          defaultMessage: 'Content Manager',
-        })}
-      />
-      <App />
-    </>
-  );
-}
