+++ packages/core/admin/admin/src/content-manager/pages/EditViewLayoutManager.tsx
@@ -98,6 +98,7 @@ const reducer = (state: EditViewState = initialState, action: Action) =>
         break;
       }
       case SET_LAYOUT: {
+        // @ts-expect-error – recursive type.
         draftState.currentLayout = action.layout;
         break;
       }
