+++ packages/core/admin/admin/src/content-manager/pages/SingleTypeRecursivePath.tsx
@@ -4,7 +4,7 @@ import { CheckPagePermissions, LoadingIndicatorPage } from '@strapi/helper-plugi
 import { Route, RouteComponentProps, Switch } from 'react-router-dom';
 
 import { useTypedSelector } from '../../core/store/hooks';
-import { useFetchContentTypeLayout } from '../hooks/useFetchContentTypeLayout';
+import { useContentTypeLayout } from '../hooks/useLayouts';
 import {
   SettingsViewComponentLayout,
   SettingsViewContentTypeLayout,
@@ -24,7 +24,7 @@ const SingleTypeRecursivePath = ({
   },
 }: SingleTypeRecursivePathProps) => {
   const permissions = useTypedSelector((state) => state.admin_app.permissions);
-  const { isLoading, layout, updateLayout } = useFetchContentTypeLayout(slug);
+  const { isLoading, layout, updateLayout } = useContentTypeLayout(slug);
 
   const { rawContentTypeLayout, rawComponentsLayouts } = React.useMemo(() => {
     let rawContentTypeLayout: SettingsViewContentTypeLayout | null = null;
