+++ packages/core/admin/admin/src/content-manager/pages/tests/NoContentType.test.tsx
@@ -1,0 +1,19 @@
+import { render, screen } from '@tests/utils';
+
+import { NoContentType } from '../NoContentType';
+
+describe('NoContentType', () => {
+  it('renders and matches the snapshot', () => {
+    render(<NoContentType />);
+
+    expect(screen.getByRole('heading', { name: 'Content' })).toBeInTheDocument();
+    expect(
+      screen.getByText(
+        "You don't have any content yet, we recommend you to create your first Content-Type."
+      )
+    ).toBeInTheDocument();
+    expect(
+      screen.getByRole('link', { name: 'Create your first Content-type' })
+    ).toBeInTheDocument();
+  });
+});
