+++ packages/core/admin/admin/src/content-manager/pages/tests/NoPermissions.test.tsx
@@ -1,0 +1,13 @@
+import { render, screen } from '@tests/utils';
+
+import { NoPermissions } from '../NoPermissions';
+
+describe('NoPermissions', () => {
+  it('renders and matches the snapshot', () => {
+    render(<NoPermissions />);
+
+    expect(
+      screen.getByText("You don't have the permissions to access that content")
+    ).toBeInTheDocument();
+  });
+});
