+++ packages/core/admin/admin/src/content-manager/utils/index.js
@@ -1,3 +1,2 @@
 export { default as arrayMoveItem } from './arrayMoveItem';
 export { default as checkIfAttributeIsDisplayable } from './checkIfAttributeIsDisplayable';
-export { default as removeKeyInObject } from './removeKeyInObject';
